+++ packages/angular-workspace/nimble-angular/label-provider/table/tests/nimble-label-provider-table.directive.spec.ts
@@ -269,22 +269,22 @@ describe('Nimble Label Provider Table', () => {
         @Component({
             template: `
                 <nimble-label-provider-table #labelProvider
-                    [cellActionMenu]="cellActionMenu"
-                    [columnHeaderGrouped]="columnHeaderGrouped"
-                    [groupCollapse]="groupCollapse"
-                    [groupExpand]="groupExpand"
-                    [rowCollapse]="rowCollapse"
-                    [rowExpand]="rowExpand"
-                    [collapseAll]="collapseAll"
-                    [columnHeaderSortedAscending]="columnHeaderSortedAscending"
-                    [columnHeaderSortedDescending]="columnHeaderSortedDescending"
-                    [selectAll]="selectAll"
-                    [groupSelectAll]="groupSelectAll"
-                    [rowSelect]="rowSelect"
-                    [rowOperationColumn]="rowOperationColumn"
-                    [rowLoading]="rowLoading"
-                    [groupRowPlaceholderNoValue]="groupRowPlaceholderNoValue"
-                    [groupRowPlaceholderEmpty]="groupRowPlaceholderEmpty"
+                    [cell-action-menu]="cellActionMenu"
+                    [column-header-grouped]="columnHeaderGrouped"
+                    [group-collapse]="groupCollapse"
+                    [group-expand]="groupExpand"
+                    [row-collapse]="rowCollapse"
+                    [row-expand]="rowExpand"
+                    [collapse-all]="collapseAll"
+                    [column-header-sorted-ascending]="columnHeaderSortedAscending"
+                    [column-header-sorted-descending]="columnHeaderSortedDescending"
+                    [select-all]="selectAll"
+                    [group-select-all]="groupSelectAll"
+                    [row-select]="rowSelect"
+                    [row-operation-column]="rowOperationColumn"
+                    [row-loading]="rowLoading"
+                    [group-row-placeholder-no-value]="groupRowPlaceholderNoValue"
+                    [group-row-placeholder-empty]="groupRowPlaceholderEmpty"
                     >
                 </nimble-label-provider-table>
             `,