+++ packages/spright-components/src/chat/input/template.ts
@@ -11,6 +11,7 @@ export const template = html<ChatInput>`
         placeholder="${x => x.placeholder}"
         rows="1"
         tabindex="${x => x.tabIndex}"
+        maxlength="${x => x.maxLength}"
         @keydown="${(x, c) => x.textAreaKeydownHandler(c.event as KeyboardEvent)}"
         @input="${x => x.textAreaInputHandler()}"
     ></textarea>